+++ affinidi-messaging-mediator/src/database/accounts.rs
@@ -1,5 +1,5 @@
 //! Handles scanning, adding and removing DID accounts from the mediator
-use affinidi_messaging_sdk::protocols::mediator::MediatorAccountList;
+use affinidi_messaging_sdk::protocols::mediator::mediator::MediatorAccountList;
 use redis::{from_redis_value, Value};
 use tracing::{debug, span, Instrument, Level};
 
@@ -85,7 +85,7 @@ impl DatabaseHandler {
     /// Retrieves up to 100 accounts from the mediator
     /// - `cursor` - The offset to start from (0 is the start)
     /// - `limit` - The maximum number of accounts to return (max 100)
-    /// NOTE: `limit` may return more than what is specified. This is a peculiarity of Redis
+    ///    NOTE: `limit` may return more than what is specified. This is a peculiarity of Redis
     pub(crate) async fn list_accounts(
         &self,
         cursor: u32,