+++ affinidi-messaging-mediator/src/messages/protocols/mediator_administration.rs
@@ -5,7 +5,7 @@ use std::time::SystemTime;
 use affinidi_messaging_didcomm::Message;
 use affinidi_messaging_sdk::{
     messages::problem_report::{ProblemReport, ProblemReportScope, ProblemReportSorter},
-    protocols::mediator::MediatorAdminRequest,
+    protocols::mediator::mediator::MediatorAdminRequest,
 };
 use serde_json::{json, Value};
 use sha256::digest;
@@ -94,12 +94,12 @@ pub(crate) async fn process(
                     }
                 }
             }
-            MediatorAdminRequest::AdminRemove(attr) => {
+            MediatorAdminRequest::AdminStrip(attr) => {
                 // Remove root admin DID in case it is in the list
                 // Protects accidentally deleting the only admin account
                 let root_admin = digest(&state.config.admin_did);
                 let attr = attr.iter().filter_map(|a| if a == &root_admin { None } else { Some(a.to_owned()) }).collect();
-                match  state.database.remove_admin_accounts(attr).await {
+                match  state.database.strip_admin_accounts(attr).await {
                     Ok(response) => {
                         _generate_response_message(&msg.id, &session.did, &state.config.mediator_did, &json!(response))
                     }