+++ affinidi-messaging-sdk/src/protocols/mediator/mod.rs
@@ -1,0 +1,3 @@
+pub mod acls;
+#[allow(clippy::module_inception)]
+pub mod mediator;